Wizard: Fix crash on "Optional Steps" in Fedora's cockpit-image-builder
With the registration step being hidden in Fedora, clicking on "Optional
Steps" leads to a crash. The reason is a bug in PatternFly, where
clicking on a parent step with a hidden first sub-step would cause
activeStep to become undefined, resulting in the crash.

The custom nav item properly handles this case by finding the first
visible sub-step instead of just the first sub-step.